+++ src/calibration_ui.cpp
@@ -76,7 +76,7 @@ auto CalibrationUi::on_spectrum_draw(const Cairo::RefPtr<Cairo::Context>& ctx) -
     auto width = allocation.get_width();
     auto height = allocation.get_height();
     auto n_bars = spectrum_mag.size();
-    auto x = util::linspace(0, width, n_bars);
+    auto x = util::linspace(0.0f, static_cast<float>(width), n_bars);
 
     for (uint n = 0u; n < n_bars - 1u; n++) {
       auto bar_height = spectrum_mag[n] * height;
